+++ examples/Snarkl/Example/Games.hs
@@ -4,9 +4,9 @@ module Snarkl.Example.Games where
 
 import Data.Typeable
 import Snarkl.Errors
-import Snarkl.Syntax
-import Snarkl.SyntaxMonad
-import Snarkl.TExpr
+import Snarkl.Language.Syntax
+import Snarkl.Language.SyntaxMonad
+import Snarkl.Language.TExpr
 import Snarkl.Toplevel
 import Prelude hiding
   ( fromRational,
