+++ examples/Snarkl/Example/Lam.hs
@@ -4,9 +4,9 @@ module Snarkl.Example.Lam where
 
 import Data.Typeable
 import Snarkl.Errors
-import Snarkl.Syntax
-import Snarkl.SyntaxMonad
-import Snarkl.TExpr
+import Snarkl.Language.Syntax
+import Snarkl.Language.SyntaxMonad
+import Snarkl.Language.TExpr
 import Prelude hiding
   ( fromRational,
     negate,
