+++ examples/Snarkl/Example/Queue.hs
@@ -6,9 +6,9 @@ import Data.Typeable
 import Snarkl.Compile
 import Snarkl.Example.List
 import Snarkl.Example.Stack
-import Snarkl.Syntax
-import Snarkl.SyntaxMonad
-import Snarkl.TExpr
+import Snarkl.Language.Syntax
+import Snarkl.Language.SyntaxMonad
+import Snarkl.Language.TExpr
 import Snarkl.Toplevel
 import Prelude hiding
   ( fromRational,
