+++ examples/Snarkl/Example/Tree.hs
@@ -3,9 +3,9 @@
 module Snarkl.Example.Tree where
 
 import Data.Typeable
-import Snarkl.Syntax
-import Snarkl.SyntaxMonad
-import Snarkl.TExpr
+import Snarkl.Language.Syntax
+import Snarkl.Language.SyntaxMonad
+import Snarkl.Language.TExpr
 import Prelude hiding
   ( fromRational,
     negate,
