+++ tests/Test/Snarkl/UnitSpec.hs
@@ -8,7 +8,7 @@ import Snarkl.Example.Lam
 import Snarkl.Example.List
 import Snarkl.Example.Peano
 import Snarkl.Example.Tree
-import Snarkl.TExpr
+import Snarkl.Language.TExpr
 import Snarkl.Toplevel
 import Test.Hspec (Spec, describe, it, shouldBe, shouldReturn)
 import Test.Snarkl.Unit.Programs
